+++ tests/framework/validators/UniqueValidatorTest.php
@@ -385,4 +385,24 @@ abstract class UniqueValidatorTest extends DatabaseTestCase
         $expected = "SELECT * FROM {$schema->quoteTableName('validator_main')} WHERE ({$schema->quoteColumnName('val_attr_b')}=:qp0) OR (val_attr_a > 0)";
         $this->assertEquals($expected, $query->createCommand()->getSql());
     }
+
+    /**
+     * Test ambiguous column name in select clause
+     * @see https://github.com/yiisoft/yii2/issues/14042
+     */
+    public function testAmbiguousColumnName()
+    {
+        $validator = new UniqueValidator([
+            'filter' => function($query) {
+                $query->joinWith('items', false);
+            },
+        ]);
+        $model = new Order();
+        $model->id = 42;
+        $model->customer_id = 1;
+        $model->total = 800;
+        $model->save(false);
+        $validator->validateAttribute($model, 'id');
+        $this->assertFalse($model->hasErrors());
+    }
 }